fixed bug with an interface as a closeable (try-with-resources statement)
JVM uses different instructions to invoke methods on an interface (INVOKEINTERFACE) and a class (INVOKEVIRTUAL). The previous implementation was considering only the second case. Such a behavior was leading to false positives on the auto-generated finally block in case of an interface as a closeable. Example #7 demonstrates the appropriate situation: https://github.com/artspb/try-with-resources-examples