Failing test for naked receiver when applied to fluent APIs
This test by @velo shows that the current implementation of the naked
receiver mutator doesn't handle the case when a chained method call gets
removed by the mutator as in the case a fluent DSLs.

The reason is that the arguments of the method call are not popped from
the stack.