Report the exception stack trace for tests that have failed (i.e. that killed a mutation). This makes it possible to check that killed mutations were killed for the expected reason (which is especially useful when there is more than one way in which a particular test can fail), and to debug cases where tests are interfering with one another (e.g. through static variables, modification of on-disk resources, etc.). The stack trace is embedded in the HTML reports. These reports are kept uncluttered (and visually identical to previously): the trace shows in a popup when clicking on a line of the Mutations section of the report. The use of a popup makes even deep stack traces very readable. Without the extra information provided by this functionality, I could not have figured out how to get PIT to give useful results on the program to which I applied it.
I implemented this change by doing some surgery on the DetectionStatus class. That class was an enumeration, which made it impossible to store the stack trace. The detection status is thus now given by an enumeration field within the DetectionStatus class, which is itself no longer an enumeration. I chose this somewhat convoluted approach to minimize changes to the rest of the code base (it was still necessary to change identity tests to equality tests in a bunch of places). This approach should be extensible and make it possible to store extra test information in different failure cases, if desired (one possibility that comes to mind is gathering stack traces in the TIMED_OUT case, to see what state the program was in).

The CSS code for the popup was adapted from http://codepen.io/imprakash/pen/GgNMXO, which is implicitly under the MIT license: http://blog.codepen.io/legal/licensing/.

Example output with stack traces is available at http://cinquin.org.uk/static/mutation_testing/uk.org.cinquin.mutinack.candidate_sequences/CandidateCounter.java.html

Signed-off-by: Olivier Cinquin <olivier.cinquin@uci.edu>