fix(report) Aggregation of heavy reports fail
In a real world project with many submodules and particularly a module that has dependencies to other modules, the generated reports can be quite heavy (100MB+). When these reports are parsed using current XML parser, report aggregation fails with GC overhead limit or OutOfMemoryError due to loading of the whole XML data in memory. The new implementation uses StAX in order to avoid loading the whole XML data in memory.