Automatically release to maven central
Release automatically when code is merged into the release branch.

The maven release plugin is no longer used, instead tracking of the
build number is moved into git, where the minor version is tracked as a tag.

Updating the major version requires manual changed to the poms, which
needs to be addressed in the future, but fortuantely major version
updates are rare.

If a major version is changed the minor version must be reset. This can be done with

```bash
git tag -d build-number-<x>
git push --delete origin build-number-<x>
```

Where `x` is the current build number.

The current development version is always x.x.0-SNAPSHOT.