Lazily initialise probe field when required
Adds a null check when accessing the probe field, initialising it if
null.

Under most circumstances the probe array will have been assigned in the
static initializer, but in the special case of a child class where the
parent class constructs the child in a static initializer the child
initializer will not have been called.