Guarantee order of mutation operators
The order of iteration of a hashset is not guaranteed. It is unclear if
this has any observable consequences, but determinism is always
preferable, and aids debugging.

The order that operators applied is now explicitly based on the mutator
id. It is unclear if there is any benefit to the de-deplication
performed by the HashSet, but this has been retained.