pluggable mutators
Mutators are now pluggable via an expanded version of the service
discovery system.

Mutator names must now match the string used to specify them on the
commandline (which is less confusing).

A new interface MutatorGroup allows mutators to be grouped together
under a name, as was previously achieved via the static calls in
Mutator.

The service loader now accepts the following types of classes

* Classes with a no arg constructor
* Enums with a single instance
* Classes with a static method named 'factory' returning a list of the
  interface

This allow most of the existing mutators to be discovered with little
modification.

Mutators can now be added by creating a plugin with an entry in a
 org.pitest.mutationtest.engine.gregor.MethodMutatorFactory file.