Mutate stream returns to Stream.empty
Previously stream returns were mutated to null. This was inconsistent
and resulted in less stable mutations. Special handling was also needed
to filter out null returns used only in flatMaps, as these acted like
stream.empty.

Mutating to Stream.empty is simpler and results in more useful mutations
that are harder to detect.