support passing project base
Although the parameter is currently unused, it would be useful in the
future to be able to identify the base dir of a project. For gradle
project this only seems to be possible from within the build script
itself.

This parameter is therefore being added without functionality so the
gradle plugin can be updated in advance.