Branch contexts with nfa state
The slot storage system in the NFA matching library was a bit of hack,
storing any matched values even if that branch do not match. Working
around this complicated using the matching library as the behaviour was
not intuitive.

This change fixes the implementation so updates to the context are tied
to the state that made them. The behaviour is now 'correct' and much
easier to understand.

Unfortunately this has two consequences.

1. It is much slower. A 1 second analysis on an example project now
   takes 8 seconds.

2. The try with resources filter is broken as it relied on peculiar
   behaviour of the original implementation.

Some mitigation has been put in place for the first point. Previously
the filters were run in the prescan (because they were so cheap). Most
filters will now not be run, unless they have the new type of
PRESCAN_FILTER.

Before this can be merged the try with resources filter must be fixed
and further benchmarks taken with large projects to understand the
impact.