select tests by block
The in memory coverage data store held a map entry for each instruction
within the code under test. This was almost unbelievably stupid and
result in very high memory consumption.

Coverage is tracked by block, which maps 1:many with instructions. The
reason coverage was being held against instruction seems to be because
the block wasn't being tracked when mutants were combined due code
inlining, so only the instruction could be used for targetting.

This change updates MutationDetails to track mutants affecting multiple
blocks, and changes the interface and internals of the coverage store
to be block driven.

For the assertj project this reduces memory consumption from 3G to 1G.
Although a large improvement, this is still too high.