poll minion processes to check they are alive
Root cause of #1047 seems to be processes that are unable to start due
to insufficient memory. This results in the main process hanging,
waiting for their signal.

At some point the checks that the minion processes are alive look to
have been removed. This change reintroduces them with a simple poll.