Support for standard argLine
Surefire and other plugins support the `argLine` property for passing
parameters to child jvms in the form of a single string.

Pitest currently supports passing a list of properties using the jvmArgs
param. This requires duplicating values already specified in the
argLine.

The list format also causes problems when the arguments contain
delimiters, for which some unsatisfactory workarounds have been applied.

This change adds a new argLine property to pitest, which is read from
the standard maven property by default. It can be set explicitly for Ant
and the command line.

The existing jvmArgs parameter is retained, and can be used to set
additional arguments for pitest.