Auto add headless true only using a feature
Pitest automatically adds -Djava.awt.headless=true to the jvm args to
prevent keyboard focu being stolen on macs.

The change introduces a new extention point ConfigurationUpdater
to allow changes to be made to the supplied config.

This interface uses the feature system, so implementations can be
enabled and disabled at runtime.

The previously hard coded adding of headless=true is moved into a
feature. It is enabled by default, but can now be dissabled by passing

features="-macos_focus"

Consideration was given to only adding the param when running on OS X,
but this was decided against and it would make the build inconsistent.