Set timeout for socket reads
Very occasionally the coverage process hangs while trying to read from
the communication socket. The root cause of this is not known, but
timing out is preferable to an infinite process hang.

Ten second timeout time is arbitary, picked to be unlikely to trigger
under normal circumstances. Set higher for mutation test minion as box
might be heavily loaded.