Capture all classloaders loading mutant
Quarkus, Roboelectric (and possibly other frameworks) load the class under test in a
classloader that is not set as the context loader when pitest starts.

A previous change got things working with Quarkus, but Roboelectric
complicates things further by persisting the classloaders between tests.
We must therefore keep track of all classloaders used to load a mutated
class, and insert each mutant into all of them.

This would have been complicated in earlier version of pitest by the
possibility that the same JVM might be used to mutate multiple classes.
This would requite the bytes of each class to be stored so that
unmutated class could be restored.

At some point, a change seems to have been made such that only 1 class
is ever mutated per jvm. Experiments to reintroduce multiple class
mutations (inners, lambdas etc) did not result in a performance increase
in sample code bases. 1 class per jvm is therefore now made a guarantee
to allow the code to be much simpler.