Relax source file attribute check
The purpose of the check is to alert users that they must configure the
compiler to include source file debug info.

Previously all classes were checked to ensure the source file was
present, however it appears there may be some scenarios in which the
source file is omitted when compiling kotlin.

To allow analysis of the remaining classes, the check has been relaxed
so that the info must be present in at least one file, rather than all
files.