use classpath jar when classpath file provided
The gradle plugin uses a classpath file to communicate with pitest when
the user classpath is too large to pass on the commandline line. In this
scenario the same issue is highly likely to be encountered when pitest
launches processes.

This change auto sets the classpath jar flag when a classpath file is
supplied to avoid the user having to set two parameters.