Auto add junit-platform-launcher to classpath
The pitest minions require a version of the junit-platform-launcher
which matches the version of junit 5 used by the SUT. Previously
this was provided by the cludge of shading it into the junit5 plugin,
but this causes constant headaches with version mismatched.

This change pulls in a matching version of platform-launcher in the same
manner that surefire does. It also allows junit dependencies to be
explicitly included pitest config should this resolution give the wrong
results.