suppress reporting of coverage stats
Arcmutates advanced incremental processing includes the option to avoid
gathering coverage for tests which cannot kill an in scope mutant.
Although this speeds up processing the html report shows most lines
as uncovered when this feature is activated. Reporting of line coverage
it therefore disabled when this feature is activated.