mutate delayed execution code stored in statics
We don't want to mutate code executed only during static initialization.
Unfortunately the current filtering also picks up code that is executed
after static initialization as lambdas.

This change implelements an imperfect comprimise where code stored as
Suppliers, Functions etc will be considered for mutation.

This will fail to mutate delayed execution code stored in other types,
and will incorrectly mutate code that is executed only during
initialization if it is within a method that returns a Supplier etc.

Although imperfect, it is an improvement.