Auto add standard kotlin source dirs if present
src/main/kotlin and src/test/kotlin are now auto added to the a maven project by the mojo if present.

This is necessary as, for maven builds, there directories are often added at runtime by the build helper or kotlin plugins. If they
pitest goal is run directly (as is common) these plugins will not fire and the config will be missing causing confusing failures.

This replaces an earlier hack where there directories were aded at a lower level.