remove use of reflection based toString generator
The reflection based toString generator fails on some modern jvms as apache commons tries
to parse out the version of Java, but fails due to outdated assumptions.

Not clear why it's inmportant to have a toString method here, but replacing with an autogenerated one
should solve #1358. Implementation excludes logger and sink form toString as it's unclear how they would be useful.