disable filters with empty method checks for interfaces
Third party plugins now support mutating annotations, so it is now
possible for mutations to be applied to interfaces. This results
in warnings in the logs from some filters that expect non
empty methods. An explicit checks for interfaces will prevent
the majority of these messages (though they may still occur
for annotations on abstract methods).